requirements.txt
speechrecognition
pyttsx3
textblob
transformers
torch
wikipedia
pillow
